fix(merchant acquirer): change type AcquirerFee to BigDecimal

+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/controller/dto/AcquirerFeeResponse.java
@@ -1,8 +1,10 @@
 package com.processing.merchantacquirer.controller.dto;
 
+import java.math.BigDecimal;
+
 public record AcquirerFeeResponse(
         String stan,
         String pan,
-        double acquirerFee
+        BigDecimal acquirerFee
 ) {
 }
