+++ services/merchant-acquirer/src/main/java/com/processing/merchantacquirer/repository/AcquirerFeeRepository.java
@@ -2,7 +2,9 @@ package com.processing.merchantacquirer.repository;
 
 import com.processing.merchantacquirer.controller.dto.AcquirerFeeRequest;
 
+import java.math.BigDecimal;
+
 public interface AcquirerFeeRepository {
-    void save(AcquirerFeeRequest key, Double amount);
-    Double get(AcquirerFeeRequest key);
+    void save(AcquirerFeeRequest key, BigDecimal amount);
+    BigDecimal get(AcquirerFeeRequest key);
 }
